Remove the editor config provider registry (#56501)
* Remove the editor_config_providers

* Remove unused translations

* Fix eslint errors


# Conflicts:
#	src/plugins/data/public/index_patterns/index_patterns/index_pattern.tsx
