+++ src/legacy/core_plugins/vis_default_editor/public/components/controls/time_interval.tsx
@@ -107,7 +107,7 @@ function TimeIntervalParamEditor({
 
   const onChange = (opts: EuiComboBoxOptionProps[]) => {
     const selectedOpt: ComboBoxOption = get(opts, '0');
-    setValue(selectedOpt ? selectedOpt.key : selectedOpt);
+    setValue(selectedOpt ? selectedOpt.key : '');
 
     if (selectedOpt) {
       agg.write();
