+++ src/plugins/data/public/index_patterns/index.ts
@@ -47,4 +47,10 @@ export const indexPatterns = {
 export { Field, FieldList, IFieldList } from './fields';
 
 // TODO: figure out how to replace IndexPatterns in get_inner_angular.
-export { IndexPattern, IndexPatterns, IndexPatternsContract } from './index_patterns';
+export {
+  IndexPattern,
+  IndexPatterns,
+  IndexPatternsContract,
+  TypeMeta,
+  AggregationRestrictions,
+} from './index_patterns';
