+++ src/plugins/data/public/index_patterns/index_patterns/index_pattern.tsx
@@ -43,6 +43,7 @@ import { formatHitProvider } from './format_hit';
 import { flattenHitWrapper } from './flatten_hit';
 import { IIndexPatternsApiClient } from './index_patterns_api_client';
 import { getNotifications, getFieldFormats, getHttp } from '../../services';
+import { TypeMeta } from './types';
 
 const MAX_ATTEMPTS_TO_RESOLVE_CONFLICTS = 3;
 const type = 'index-pattern';
@@ -54,7 +55,7 @@ export class IndexPattern implements IIndexPattern {
   public title: string = '';
   public type?: string;
   public fieldFormatMap: any;
-  public typeMeta: any;
+  public typeMeta?: TypeMeta;
   public fields: IFieldList;
   public timeFieldName: string | undefined;
   public intervalName: string | undefined | null;
@@ -423,6 +424,10 @@ export class IndexPattern implements IIndexPattern {
   getFieldByName(name: string): Field | void {
     if (!this.fields || !this.fields.getByName) return;
     return this.fields.getByName(name);
+  }
+
+  getAggregationRestrictions() {
+    return this.typeMeta?.aggs;
   }
 
   isWildcard() {
