+++ src/plugins/data/public/index_patterns/index_patterns/types.ts
@@ -17,16 +17,19 @@
  * under the License.
  */
 
-import { VisState } from 'src/legacy/core_plugins/visualizations/public';
-import { IAggConfig, AggParam } from '../../legacy_imports';
-import { EditorConfig } from '../utils';
+export type AggregationRestrictions = Record<
+  string,
+  {
+    agg?: string;
+    interval?: number;
+    fixed_interval?: string;
+    calendar_interval?: string;
+    delay?: string;
+    time_zone?: string;
+  }
+>;
 
-export const aggParamCommonPropsMock = {
-  agg: {} as IAggConfig,
-  aggParam: {} as AggParam,
-  editorConfig: {} as EditorConfig,
-  formIsTouched: false,
-  metricAggs: [] as IAggConfig[],
-  state: {} as VisState,
-  showValidation: false,
-};
+export interface TypeMeta {
+  aggs?: Record<string, AggregationRestrictions>;
+  [key: string]: any;
+}
