+++ x-pack/legacy/plugins/lens/public/indexpattern_plugin/types.ts
@@ -5,6 +5,7 @@
  */
 
 import { IndexPatternColumn } from './operations';
+import { AggregationRestrictions } from '../../../../../../src/plugins/data/public';
 
 export interface IndexPattern {
   id: string;
@@ -19,18 +20,6 @@ export interface IndexPattern {
     }
   >;
 }
-
-export type AggregationRestrictions = Record<
-  string,
-  {
-    agg?: string;
-    interval?: number;
-    fixed_interval?: string;
-    calendar_interval?: string;
-    delay?: string;
-    time_zone?: string;
-  }
->;
 
 export interface IndexPatternField {
   name: string;
