+++ x-pack/plugins/translations/translations/ja-JP.json
@@ -10394,8 +10394,6 @@
     "xpack.rollupJobs.detailPanel.jobActionMenu.buttonLabel": "管理",
     "xpack.rollupJobs.detailPanel.loadingLabel": "ロールアップジョブを読み込み中…",
     "xpack.rollupJobs.detailPanel.notFoundLabel": "ロールアップジョブが見つかりません",
-    "xpack.rollupJobs.editorConfig.dateHistogram.customInterval.helpText": "ロールアップ構成の間隔の倍数でなければなりません: {interval}",
-    "xpack.rollupJobs.editorConfig.histogram.interval.helpText": "ロールアップ構成の間隔の倍数でなければなりません: {interval}",
     "xpack.rollupJobs.editRollupIndexPattern.createIndex.defaultButtonDescription": "要約データに制限された集約を実行します。",
     "xpack.rollupJobs.editRollupIndexPattern.createIndex.defaultButtonText": "ロールアップインデックスパターン",
     "xpack.rollupJobs.editRollupIndexPattern.createIndex.defaultTypeName": "ロールアップインデックスパターン",
